fix: remove night greeting bucket and extract shared useGreeting hook
Per maintainer feedback, "Good night" is a farewell not a greeting.
Simplifies to three buckets: morning (5-11), afternoon (12-16),
evening (17-4).

Extracts duplicated timezone validation, Intl formatting, and daypart
logic into a shared useGreeting hook to eliminate future drift between
the two identical components.

Removes greetings.night key from all 19 locale files.
