+++ packages/i18n/src/locales/cs/translations.ts
@@ -494,7 +494,6 @@ export default {
     morning: "Dobré ráno, {first_name} {last_name}",
     afternoon: "Dobré odpoledne, {first_name} {last_name}",
     evening: "Dobrý večer, {first_name} {last_name}",
-    night: "Dobrou noc, {first_name} {last_name}",
   },
   show_all: "Zobrazit vše",
   show_less: "Zobrazit méně",
