+++ packages/i18n/src/locales/de/translations.ts
@@ -507,7 +507,6 @@ export default {
     morning: "Guten Morgen, {first_name} {last_name}",
     afternoon: "Guten Nachmittag, {first_name} {last_name}",
     evening: "Guten Abend, {first_name} {last_name}",
-    night: "Gute Nacht, {first_name} {last_name}",
   },
   show_all: "Alle anzeigen",
   show_less: "Weniger anzeigen",
