+++ packages/i18n/src/locales/en/translations.ts
@@ -327,7 +327,6 @@ export default {
     morning: "Good morning, {first_name} {last_name}",
     afternoon: "Good afternoon, {first_name} {last_name}",
     evening: "Good evening, {first_name} {last_name}",
-    night: "Good night, {first_name} {last_name}",
   },
   show_all: "Show all",
   show_less: "Show less",
