+++ packages/i18n/src/locales/ro/translations.ts
@@ -500,7 +500,6 @@ export default {
     morning: "Bună dimineața, {first_name} {last_name}",
     afternoon: "Bună ziua, {first_name} {last_name}",
     evening: "Bună seara, {first_name} {last_name}",
-    night: "Noapte bună, {first_name} {last_name}",
   },
   show_all: "Arată tot",
   show_less: "Arată mai puțin",
