+++ packages/i18n/src/locales/ru/translations.ts
@@ -502,7 +502,6 @@ export default {
     morning: "Доброе утро, {first_name} {last_name}",
     afternoon: "Добрый день, {first_name} {last_name}",
     evening: "Добрый вечер, {first_name} {last_name}",
-    night: "Доброй ночи, {first_name} {last_name}",
   },
   show_all: "Показать все",
   show_less: "Свернуть",
