+++ packages/i18n/src/locales/sk/translations.ts
@@ -494,7 +494,6 @@ export default {
     morning: "Dobré ráno, {first_name} {last_name}",
     afternoon: "Dobré popoludnie, {first_name} {last_name}",
     evening: "Dobrý večer, {first_name} {last_name}",
-    night: "Dobrú noc, {first_name} {last_name}",
   },
   show_all: "Zobraziť všetko",
   show_less: "Zobraziť menej",
