+++ packages/i18n/src/locales/ua/translations.ts
@@ -502,7 +502,6 @@ export default {
     morning: "Добрий ранок, {first_name} {last_name}",
     afternoon: "Добрий день, {first_name} {last_name}",
     evening: "Добрий вечір, {first_name} {last_name}",
-    night: "Доброї ночі, {first_name} {last_name}",
   },
   show_all: "Показати все",
   show_less: "Показати менше",
