+++ packages/i18n/src/locales/vi-VN/translations.ts
@@ -500,7 +500,6 @@ export default {
     morning: "Chào buổi sáng, {first_name} {last_name}",
     afternoon: "Chào buổi chiều, {first_name} {last_name}",
     evening: "Chào buổi tối, {first_name} {last_name}",
-    night: "Chúc ngủ ngon, {first_name} {last_name}",
   },
   show_all: "Hiển thị tất cả",
   show_less: "Hiển thị ít hơn",
