+++ packages/i18n/src/locales/zh-TW/translations.ts
@@ -481,7 +481,6 @@ export default {
     morning: "早安，{first_name} {last_name}",
     afternoon: "午安，{first_name} {last_name}",
     evening: "晚上好，{first_name} {last_name}",
-    night: "晚安，{first_name} {last_name}",
   },
   show_all: "顯示全部",
   show_less: "顯示較少",
